+++ CHANGELOG.rst
@@ -6,6 +6,7 @@ Changelog for package layered_hardware_gazebo
 --------------------
 * Add mimic/mimic-PID operation modes
 * Supress a warning on unloading layer plugins
+* Initialize joint states on starting a GazeboJointLayer
 
 0.0.4 (2020-05-05)
 ------------------